Compress instead of serialize data when loading a dupe
If we `util.Compress` the data, we send significantly less of it over the data. The again fixes the serialize error caused by trying to send more than 64kb of data, but only because the `serialize` library wasn't really compressing it. This doesn't solve the problem when the `util.Compress` data size is >64kb, but it should be much harder for the limit to be reached. I'll write another commit to send the data in chunks again.